Initial NR numerology implementation in MmWaveEnbPhy
This PR reflect the ones introduced in "Initial NR numerology implementation in MmWaveUePhy". Specifcally, the ENB now scans through TTIs (instead of slots), eventually ending and starting new slots (instead of subframes).

In addition to the aforementioned changes, which are similar to the UE case, we have the management of MIBs and SIBs which (at least for now) mimics the previously implemented behaviour, but with a different periodicity of 20ms; with MIBs transmitted on even frames while SIBs on odd ones.