Prevent getting old BF vector if initial conf in mmwave-CbBf (#91)
Currently, the method ComputeBeamformingCodebookMatrix of the class MmWaveCodebookBeamforming stores the previous beamforming (BF) vectors before computing the matrix associating gNB and UE codebook indexed and the corresponding SNR values. However, this leads to a crash in the first configuration, since the phased arrays do not have any BF vector yet.
This PR aims at fixing this bug by storing the BF vectors only for the configurations which follow the initial one.