Memory optimization for energy efficiency heuristics (#117)
Refactoring of the code to prevent the leaking

lte: Allocate non-contention-based preamble only if it's free

The non-contention preamble assigned to UE during handover or PDCCH order is valid only until the time duration of the “expiryTime” of the preamble is reached. This timer value is only maintained at the eNodeB and the UE has no way of knowing if this timer has expired. If the UE tries to send the preamble again after the expiryTime and the preamble is re-assigned to another UE, it results in errors. This has been solved by re-assigning the preamble to another UE only if it is not being used (An UE can be using the preamble even after the expiryTime duration).

Commits:
- https://gitlab.com/nsnam/ns-3-dev/-/commit/57160aa7e20eb2e33232673f9a6fe7c807534265
- https://gitlab.com/nsnam/ns-3-dev/-/commit/8137af04f0c60c19cd6e53fb7b2ba60b6e7fdf78#e71b3917243ad08b8a32a88191e34c4147f381ea